Use basedpyright for type checking
basedpyright has docstrings in its builtins and standard library type
stubs, so it is significantly better than pyright. But it is also too
dumb to be opinionated: For example, it can't infer member type in
`x = []; x.append(2)` and complains. (If it is because of python being
too dynamic, checks that do not work and are still enabled by default is
stupid.)